Feat(web-react): Abstract shared props between Text and Heading into shared type #DS-2127
